+++ test/evaluate.test.js
@@ -58,4 +58,14 @@ describe('Basic parsing', () => {
     let data = await value.get()
     expect(data).toStrictEqual(["Michael"])
   })
+
+  test("Non-array documents", async () => {
+    let dataset = {data: [{person: {_ref: 'b'}}]}
+
+    let query = `*.data{person->}`
+    let tree = parse(query)
+    let value = await evaluate(tree, {dataset})
+    let data = await value.get()
+    expect(data).toStrictEqual([{}])
+  })
 })